requirements.txt
Flask
flask-swagger-ui
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
Flask-Limiter
Flask-Caching
python-jose
Werkzeug